RH: add mpathconf
mpathconf is a program (largely based on lvmcomf) to help users
configure /etc/multipath.conf and enable or disable multipathing.  It
has a couple of built-in options that can be set directly from the
command line.  But, mostly it is used to get a multipath.conf file
with the OS defaults, and to enable and disable multipathing via
a single command.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>