multipathd: don't access path if it was deleted
ev_remove_path() could fail and still delete the path. This could cause
problems for handle_path_wwid_change(), which expected that a failure
meant that the path still existed. ev_remove_path now returns a
different error code for failure to reload the multipath device, so that
it can be differentiated from cases where the path was no removed.