Revert "multipath-tools: make EMC/Invista config work with alua and multibus"
This reverts commit 309ff281aaa07e862540d3d645a8263f3e9baaed.

Mail from <Wayne.Berthiaume@dell.com>, 20210930:

"OPM is no longer supported in the Dell VPLEX product. If we at Dell had
wished to change the default device stanzas for any of our products they
would have been done when the product and/or feature is released.
Please remove this patch as well. It is not needed."