libmpathpersist: update reservation key before checking paths
There is a race condition when changing reservation keys where a failed
path could come back online after libmpathpersist checks the paths, but
before it updates the reservation key. In this case, the path would come
up and get reregistered with the old key by multipathd, and
libmpathpersist would not update its key, because the path was down
when it checked.

To fix this, check the paths after updating the key, so any path that
comes up after getting checked will use the updated key.

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>